refactor: break the game into components
