+++ src/modules/GameImaginarium/components/Dashboard/index.ts
@@ -1,0 +1,1 @@
+export { Dashboard } from './Dashboard';
