+++ src/modules/GameImaginarium/components/WordImage/index.ts
@@ -1,0 +1,1 @@
+export { WordImage } from './WordImage';
