+++ src/modules/GameImaginarium/components/index.ts
@@ -1,0 +1,2 @@
+export { WordImage } from './WordImage';
+export { Dashboard } from './Dashboard';
